Extract Component ArticleWrapperHeader
To make component smaller.
